// cargodelivery-front/generated/models/Car.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { CarColor } from './CarColor';
import {
    CarColorFromJSON,
    CarColorFromJSONTyped,
    CarColorToJSON,
} from './CarColor';

/**
 * 
 * @export
 * @interface Car
 */
export interface Car {
    /**
     * 
     * @type {string}
     * @memberof Car
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Car
     */
    mark: string;
    /**
     * 
     * @type {string}
     * @memberof Car
     */
    model: string;
    /**
     * 
     * @type {string}
     * @memberof Car
     */
    plateNumber: string;
    /**
     * 
     * @type {CarColor}
     * @memberof Car
     */
    color: CarColor;
}

/**
 * Check if a given object implements the Car interface.
 */
export function instanceOfCar(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "mark" in value;
    isInstance = isInstance && "model" in value;
    isInstance = isInstance && "plateNumber" in value;
    isInstance = isInstance && "color" in value;

    return isInstance;
}

export function CarFromJSON(json: any): Car {
    return CarFromJSONTyped(json, false);
}

export function CarFromJSONTyped(json: any, ignoreDiscriminator: boolean): Car {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'mark': json['mark'],
        'model': json['model'],
        'plateNumber': json['plateNumber'],
        'color': CarColorFromJSON(json['color']),
    };
}

export function CarToJSON(value?: Car | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'mark': value.mark,
        'model': value.model,
        'plateNumber': value.plateNumber,
        'color': CarColorToJSON(value.color),
    };
}